update amount data type to BigDecimal instead of double to prevent rounding errors

+++ src/main/java/com/example/payment/EmailService.java
@@ -1,5 +1,7 @@
 package com.example.payment;
 
+import java.math.BigDecimal;
+
 public interface EmailService {
-    void sendPaymentConfirmation(String email, double amount);
+    void sendPaymentConfirmation(String email, BigDecimal amount);
 }
