+++ src/main/java/com/example/payment/PaymentApi.java
@@ -1,5 +1,7 @@
 package com.example.payment;
 
+import java.math.BigDecimal;
+
 public interface PaymentApi {
-    PaymentApiResponse charge(String apiKey, double amount);
+    PaymentApiResponse charge(String apiKey, BigDecimal amount);
 }
